Tweaked minimum RD values.
Added tracking of summary statistics for kill weapons.

Changed the calculation to be chunk size based instead of expecting a time based calculation for chunks.  This makes the calculation 100% reproducible on restarts.

Removed overall and noarmr ranks from the default output.  Will add a flag to include those at some point in the future.  The calculations are still done, just not sent by default.